Enable tricircle commands
1. What is the problem?
There is no direct way to call tricircle commands.  In consequence,
every command has to be consumed thru python interpreter.
  e. g. python cmd/manage.py etc/api.conf

2. What is the solution to the problem?
The solution is to list the new commands in the pbr configuration
file. The pdr configuration requires a valid python modules, but
given that they have implementated in *cmd* directory(instead of
*tricircle/cmd*) is required to move them. As initial phase
commands will be copied, once there is no reference to the old ones
it's possible to proceed to deprecate them.

3. What the features need to be implemented to the Tricircle to
realize the solution?
Replace calls to duplicate module and deprecate the old one.

Change-Id: I6a27d990803e928151ca424f47564b6626c8e99b