spec on legacy tables clean work
1. What is the problem?
After the Tricircle narrowed its scope to networking automation across Neutron
servers, many table are deprecated, therefore they should be removed. Other
tables that still exist in the Tricircle need to be renamed for better
understanding.

2. What is the solution to the problem?
The detailed description of the problems and the proposals to solve the
problems will be covered in this specification.

3. What the features need to be implemented to the Tricircle to realize
the solution?
The new features discussed in this specification needs to be implemented.

Change-Id: Idf61ecb30f54e40fde283ee5f3278b2f4b429f66