Add release notes for resource routing api
1. What is the problem?
Both the blueprint and implementation of the resource routing
API have been completed, so release notes are required for
these new features. This patch adds the release notes for them.
Moreover, the releasenotes test environment should be added
to the tox.ini. The reno version should be given in
test-requirements.txt.

2. What is the solution to the problem?
Add release notes for the resource routing API in the new
file releasenotes/notes/resource-routing-operation.yaml.
Add the releasenotes test environment to file tox.ini and reno
version to the test-requirements.txt.

3. What the features need to be implemented to the Tricircle
to realize the solution?
No new features.

Change-Id: I23c9b3fb97b33f15a877eb2dbcabbb823c825d6d