Fix security group rule sync
1. What is the problem
After security group rule sync-ed by XJob daemon, IPv4 ingress rule
included in the DEFAULT security group is missing.

2. What is the solution to the problem
After checking the code, we find that the indenation in XManager is
incorrect so XJob daemon doesn't collect the correct list of rules
needed to be sync-ed to the local Neutron server. So the solution is
simple, we just correct the indentation.

3. What the features need to be implemented to the Tricircle
No new features

Change-Id: I981add46eb8fee8f1c025de0ba1805a8732f6cc5
Closes-Bug: #1655271