Change network type shared_vlan name to vlan
1. What is the problem
shared_vlan's concept is to create a network spanning into multiple
OpenStack with same vlan segment, but when create a network which only
resides in one region and the network type is vlan, we have to specify
the physical network type as shared_vlan, otherwise no other way to
create a vlan network in local Neutron.

2. What is the solution to the problem
Just use vlan as the physical network type, and use
availability_zone_hints to limit where the network will reside.

3. What the features need to be implemented to the Tricircle
No new features

Change-Id: Ib3f110e2281eff2997752debda319da282c3e3ad