Fix Xjob of router setup not triggered
1. What is the problem?
When booting instance and adding router interface are running at
the same time, sometimes asynchronous job of router setup is not
triggered so the bottom router is not created. A script that can
reproduce this problem is attached to the bug report page[1].

2. What is the solution to the problem?
As discussed in the bug report page[1], this is a time series
related bug. To fix this bug, we just need to modify the router
interface adding process to add top interface before getting
network resource routing entries.

3. What the features need to be implemented to the Tricircle
to realize the solution?
No new feature.

[1] https://bugs.launchpad.net/tricircle/+bug/1647924

Change-Id: Ia79e375ed95fc88f5f670e350fe5c7bbef5627f2
Closes-Bug: #1647924