Fix unit tests error introduced by _get_session
1. What is the problem
In neutron.plugins.ml2.drivers.helpers._get_session, the passed-in
argument is first checked if it's a neutron.context.Context object,
the argument will be directly returned if it's not a neutron Context
object. In our test cases, the FakeNeutronContext does not inherit
neutron Context object, so the session inside FakeNeutronContext is
not returned.

2. What is the solution to the problem
Make FakeNeutronContext inherit neutron Context object. A bump_revision
method is also added to DotDict object since Neutron starts to call
this method.

3. What the features need to be implemented to the Tricircle
No new features

Change-Id: Ib17eb96418088217e991e52501209e0fc1227213