Cross-pod VxLAN network spec
1. What is the problem?
VLAN network has some restrictions that VxLAN network doesn't have.
For more flexible networking deployment, we consider supporting
cross-pod VxLAN network.

2. What is the solution to the problem?
The detailed discussion of the proposals to support cross-pod VxLAN
network is covered in this specification.

3. What features need to be implemented to the Tricircle
to realize the solution?
The proposed changes in this specification need to be implemented.

Change-Id: I79b41a63778660bd5a7b6e5e3d66ce0bd394e9af