Catch an exception on resource routing update
1. What is the problem
When resource routing needs to update its attribute
named pod_id, it will fail if the pod where the new pod_id
lays on doesn't exist. This update error was caught
through internal server error originally and it was not
detailed enough.

2. What is the solution for the problem
Dismiss this update request by giving a regular invalid
parameter error and tell the operator that the new pod
for resource routing should exist. The pod to which the
pod_id belongs can only be selected from the existing
pods. Because it is a foreign key and highly depends on
pod table. At the same time, we should add one test case
for this circumstance.

3. What the features need to be implemented to the Tricircle to
realize the solution
None.

Change-Id: Iae6e3368c7f0ff0cd751aca932cba818328bb723