spec on asynchronous job management
1. What is the problem
When XJob receives a job message from service, it will register
the job in database and handle it asynchronously. Tricircle
need to provide API for users to query the job status and trigger
failed job if something happens unexpectedly. This patch provides
asynchronous job management specification, then we will develop
relevant API to manage the job according to this specification.

2. What is the solution for the problem
The new features functioning as job management APIs are covered
in this specification.

3. What the features need to be implemented to the Tricircle to
realize the solution
These asynchronous job management APIs in this specification need
to be implemented.

Change-Id: I58df65b042a8866c4d8d7f5a65a3f85bcf43b62b