Add multi-region configuration in gate_hook.sh
1. What is the problem
Currently the function test job in Tricircle can only
work in single region, and most of multi-region test cases
are not able to be tested without multi-region test
environment.

2. What is the solution for the problem
Add multi-node gate job, and configure the subnode to the second
region, and use the keystone in the primary node, so that
multi-region gate/check test environment could be established
for Tricircle

3. What the features need to be implemented to the Tricircle to
realize the solution
No feature delivered to end user, just enable multi-region test.

Change-Id: I9f3d50b4fcf7796f39d3342d9a044d359f4fd7e9
Signed-off-by: joehuang <joehuang@huawei.com>