Networking QoS service spec
1. What is the problem
QoS is defined as the ability to guarantee certain network requirements
like bandwidth, latency, jitter, and reliability in order to satisfy a
Service Level Agreement (SLA) between an application provider and end
tenants. Networks or ports in local Neutron can not be associated with
QoS policy automatically yet.

2. What is the solution to the problem
Using asynchronous jobs to complete the automation of QoS policy/rule
in local Neutrons.

3. What the features need to be implemented to the Tricircle
to realize the solution?
QoS.

Change-Id: I8dcce7d35d7fe7283cdeb80d71259d3f3859406b