Implement asynchronous job Admin API
1. What is the problem
When XJob receives a job message from service, it will register
the job in database and handle it asynchronously. Tricircle
needs to provide API for admin to query the job status and trigger
failed job if something happens unexpectedly. The detailed work
for XJob Admin APIs is covered in the document[1].

2. What is the solution for the problem
We implement XJob management APIs, they are listed as following:
 *(1) create a job
 *(2) list single job info
 *(3) list all jobs
 *(4) list jobs with filters
 *(5) list all jobs' schemas
 *(6) delete a job
 *(7) redo a job

3. What the features need to be implemented to the Tricircle to
realize the solution
Implement above job operations.

[1] https://review.openstack.org/#/c/438304/

Change-Id: Ibd90e539c9360a0ad7a01eeef185c0dbbee9bb4e