Improve operation of creating ports
1. What is the problem?
When creating a load balancer, the service plugin creates a port for the
load balancer, whose fixed_ips is filled with subnet_id. As a result, it
is processed as ip specified cases by mistake.

2. What is the solution to the problem?
To ensure the operation of creating port can be processed as ordinary
cases, extra judgement for load balancer's port is added.

3.  What features need to be implemented to the Tricircle to
realize the solution?
No new features.

Change-Id: If70ca6258d6c518c95e060691d2547d9bbf7359d