Mark subnet deletion state during subnet deletion
1. What is the problem
As reported in the bug page, local subnet will be recreated
after central subnet is deleted

2. What is the solution for the problem
The reason for the recreation is that DhcpAgentNotifyAPI subscribe
the subnet deletion event and will call get_network during subnet
deletion, get_network will call _ensure_subnet to create the subnet.
So we need a way to distinguish the call from API and callback.

A dict on_subnet_delete is added to local plugin in this patch.
In delete_subnet, before calling real core pluing's delete_subnet
method, we first add a new item (request_id, True) to this dict.
After the calling, this item is removed. So in get_network, we can
check whether the dict contains the request_id, if so, we know
that we are in the subnet deletion stage, then we don't call
_ensure_subnet.

3. What features need to be implemented to the Tricircle to
realize the solution
No new features

Change-Id: Ie79460aa763918be7c522263d9865847fc0da828
Closes-Bug: #1680757