Support multi-project in xjob
1. What is the problem
We utilize an admin user to operate resources in asynchronous job.
Currently the project of this user is pre-configured, so when this
user try to operate resources which do not belong to the configured
project, operation will fail.

2. What is the solution for the problem
During deployment, we create an admin user and assign admin role to
it in each projects that need to be operated. And in the codes, since
now we have project_id in job table, we can use this project_id to
retrieve an admin token, instead of using the pre-configured project_id.

3. What features need to be implemented to the Tricircle to
realize the solution
Support using different project_id according to the operated resources
to get admin token when running asynchronous job.

Change-Id: Id3eb464bc0573f96cecfa4700cc977e328054d10