Don't reset endpoint_url in ResourceHandle but Client
1. What is the problem
In resource_handle.NeutronResourceHandle, after connection-fail
exception is caught, endpoint_url attribute is set to None and waits
for client.Client to update the endpoint.

However, local plugin directly uses resource_handle to talk to central
Neutron, so the endpoint_url attribute will not be updated after it's
set to None. Actually, for local plugin, endpoint_url should be read
from the configuration file and doesn't require updating, so it's not
necessary to set it to None.

2. What is the solution for the problem
Reset endpoint_url in Client instead of ResourceHandle

3. What features need to be implemented to the Tricircle to
realize the solution
N/A

Change-Id: If306031605f30adf933464b955386162c7f83c06
Closes-Bug: #1691306