Local FIP not created after central FIP creation
1. What is the problem
After booting a VM, we associate a FIP to it via central Neutron.
However, local FIP is not created as expected. Also, there is no related
asynchronous job registered.

The reason may be that we don't trigger asynchronous job in
"create_floatingip" method, and since VM is already there, there's no
chance to trigger the job.

2. What is the solution for the problem
If "port_id" is specified in the body, trigger asynchronous job.

3. What features need to be implemented to the Tricircle to
realize the solution
Support creating FIP with "port_id" specified.

Change-Id: Iceb15f68acc23f8dcb8767ac0371947334a1e9db
Closes-Bug: #1691918