Work with cell v2
1. What is the problem
We try to figure out a way to integrate Tricircle with Nova cell v2.

2. What is the solution for the problem
The basic idea is to start a local Neutron server for each cell. Nova-
compute in each cell is configured to talk to local Neutron server in
the same cell. All local Neutron servers are configured to talk to the
very one Nova-API.

Currently DevStack doesn't support multi-cell deployment, so we try to
deploy it with our own plugin. In node1, Nova services are started as
before, but we change the region of Nova-API, Glance-API and placement-API
from RegionOne to CentralRegion. Local Neutron server is also configured
to talk to Nova-API in CentralRegion. In node2, Nova-API is enabled at
first, since we need DevStack to help us create the related database.
After the DevStack start in the node2, we manually stop Nova-API and
Nova-scheduler in node2.

One document discussing the detailed setup and trouble shooting steps
is added.

3. What features need to be implemented to the Tricircle to
realize the solution
Tricircle can work with Nova cell v2.

Change-Id: I6ba8e1022d83f40df36464abfdd7b4844673b24d