Boot a VM in a subnet without gateway
1. What is the problem
We can create a subnet without gateway in the central Neutron, but when
we boot a VM in that subnet, the request fails since local Neutron is
trying to calculate the IP allocation pool based on the central subnet
gateway IP, which is None.

2. What is the solution for the problem
If a central subnet doesn't have gateway, do not re-calculate local IP
allocation pool when creating local subnet.

3. What features need to be implemented to the Tricircle to
realize the solution
Support booting a VM in a subnet that has not gateway

Change-Id: I04f7f9f9926f22bca3490ee46e2163b9c17bb8fe
Closes-Bug: #1693446