Router update should not touch nonexistent local router
1. What is the problem
Set external gateway to a router, if error occurs during local router
creating/updating, exception will be raised, but this time central
router is already updated with external gateway information. So when we
try to unset the external gateway, central Neutron server will try to
remove external gateway for local router first, if local router doesn't
exist at this time, we have no way to unset the external gateway since
we fail to remove external gateway for local router.

2. What is the solution for the problem
Do not try to remove external gateway for a nonexistent local router.

3. What features need to be implemented to the Tricircle to
realize the solution
N/A

Change-Id: Id7261640405d4d24c8523f6518ec19edc031024b
Closes-Bug: #1693138