Support pagination for resource routing list operations
1. What is the problem
The resource routing list operations will retrieve all the items in
the database, which will consume too much memory and take long time
to response when the results are considerably large.

2. What is the solution for the problem
To reduce load on the service, list operations will return a maximum
number of items at a time by pagination. To navigate the collection,
the parameters limit and marker can be set in the URI. For example:

/v1.0/routings?limit=2000&marker=500

The marker parameter is the ID of the last item in the previous list.
A marker with an invalid ID returns a badRequest (400) fault.
The limit parameter sets the page size. These parameters are optional.
If the client requests a limit beyond the maximum limit configured by
the deployment, the server returns the maximum limit number of items.
Pagination and filtering can work together with routing's list operations.

3. What the features need to be implemented to the Tricircle to
realize the solution
Add pagination feature for resource routing list operations.

Change-Id: I05d1b30f502103d247d8be06c1e52fdcec42b41e