[Urgent] Fix fip creation error
1. What is the problem
In the patch[1] for solving bug[2], we introduce an error that when
a fip is created with port ID and the bottom port is not there, the
fip body is not returned. Our smoke test doesn't find out this error
because it only affects the response, the fip is correctly created
in the database. But the log does show an error message.

2. What is the solution for the problem
Always return fip body. Unit test is modified to use the returned
body to ensure it's not None.

3. What features need to be implemented to the Tricircle to
realize the solution
N/A

[1] https://review.openstack.org/#/c/465870
[2] https://launchpad.net/bugs/1691918

Change-Id: I71fcc8fab9fb2a50ab058d67980f6e1e1e4e6d90