Spec for smoke test engine
1. What is the problem
Problems in the current smoke test:
(1) Mix use of bash and python scripts
(2) Resources are not cleaned at the end of the test

2. What is the solution for the problem
A proposed solution is discussed in the spec. The basic idea is
to define tests using YAML and use a engine to parse YAML and
run tests.

3. What features need to be implemented to the Tricircle to
realize the solution
A task engine and a task runner.

Change-Id: Ib7ca2242529e147e9edd6292eec3003967ea83f0