Fix bugs in tricircle Admin API
1. What is the problem
When we use admin role for demo project to retrieve jobs/routings
through Admin API, return results include admin project's
jobs/routings. The result should be only related to demo project.

2. What is the solution for the problem
Extract project ID from context and use it as filter. The project
ID filter in URL query string will be ignored, and only the project ID
in which the user is authorized will be used as the filter.

3. What the features need to be implemented to the Tricircle to
realize the solution
Choose project ID as mandatory filter for jobs/routings list operation.

Change-Id: I86778e0525c1fecf45bc473bd42e62909534778b
Closes-Bug: #1711001
Closes-Bug: #1711003