Distinguish the source of requests
1. What is the problem?
When receiving a request, we don't know whether the request
comes from  'Central Neutron' or 'Local Neutron'.

2. What is the solution to the problem?
In order to determine the source of requests, we add a
filter to get the source message in the header of requests.
Next step we tag the source message into the context. When
deploy the WSGI app, we checkout the User-Agent in the
request's headers and tag the requests.

3. What the features need to be implemented to the Tricircle
to realize the solution?
No

Change-Id: I990fa46e887cc0822b8e6d74d199d92e39df0bd6