Fix l3 unit test error
1. What is the problem?
Our l3 unit test fails because of a patch[1] that migrates to use OVO
model for subnet and the model validates whether the ipv6_ra_mode and
ipv6_address_mode fields are set as the enumeration values.

2. What is the solution to the problem?
Set ipv6_address_mode and ipv6_ra_mode as "slaac" to meet the requirement.

3. What features need to be implemented to the Tricircle to realize
the solution?
N/A

[1] https://github.com/openstack/neutron/commit/af00cbc3d248e7beec3e6e49dda963051b85d496

Change-Id: Icaaaf7fc08eb689dbf4437da9fe0714fcdedc250