spec for resource deleting
1. What is the problem?
During the deletion of a network which is mapped to several local networks, if
the local neutron server receives a network-get request, it will bring some
conflict operations. For example, we delete local networks before deleting
central network. If a "get_network" request comes to a local neutron server
after the local network is completely deleted in that region, if central
network is still there (assuming it takes certain time to delete all local
networks), Tricircle will also query central neutron and the deleted local
network will be recreated.

2. What is the solution to the problem?
The solution to resource deleting is covered in this specification.

3. What the features to be implemented in the Tricircle
to realize the solution?
No new features.

Change-Id: I4f29ee9f214242262bffc616dc60c8fdde76a6f6
Signed-off-by: song baisen <songbaisen@szzt.com.cn>
Co-Authored-By: CR_hui <CR_hui@126.com>