Modify code for release port when delete vm
1. What is the problem?

In a non-local network, delete a vm from one reigon, and use its port to
create a vm in another region, the port in the original region is not
released.

2. What is the solution to the problem?

Release port to down when delete a vm from one region, and change shadow
port 'name' and 'device_id' and 'device_owner' value if shadow port exist.
In the update port, according to the update before the port and the
updated port to determine whether to notify the Central updates.

3. What the features to be implemented in the Tricircle to realize the solution?

Because port be cached in neutron-openvswitch-agent, not delete port
when delete a vm from one region.

Signed-off-by: zhang xiaohan <zhangxiaohan@szzt.com.cn>
Co-Authored-By: tangzhuo <ztang@hnu.edu.cn>
Change-Id: If6819bf67ea2d206cc735039d8c55733dae54ee9